requirements.txt
discord.py==0.16.12
python-dotenv==0.9.1
beautifulsoup4==4.6.3
asyncio==3.4.3
